# Optimize program size by refactor error reporting routines