--- a/src/V3Error.h
+++ b/src/V3Error.h
@@ -286,9 +286,12 @@ inline std::ostream& operator<<(std::ostream& os, const V3ErrorCode& rhs) {
 }
 
 // ######################################################################
+class V3Error;
+
 class V3ErrorGuarded final {
     // Should only be used by V3ErrorGuarded::m_mutex is already locked
     // contains guarded members
+    friend class V3Error;
 public:
     using MessagesSet = std::set<std::string>;
     using ErrorExitCb = void (*)(void);
@@ -319,6 +322,16 @@ private:
         = MAX_ERRORS;  // Option: --error-limit Number of errors before exit
     bool m_warnFatal VL_GUARDED_BY(m_mutex) = true;  // Option: --warnFatal Warnings are fatal
     std::ostringstream m_errorStr VL_GUARDED_BY(m_mutex);  // Error string being formed
+
+    void v3errorPrep(V3ErrorCode code) VL_REQUIRES(m_mutex) {
+        m_errorStr.str("");
+        m_errorCode = code;
+        m_errorContexted = false;
+        m_errorSuppressed = false;
+    }
+    std::ostringstream& v3errorStr() VL_REQUIRES(m_mutex) { return m_errorStr; }
+    void v3errorEnd(std::ostringstream& sstr, const string& extra = "") VL_REQUIRES(m_mutex);
+
 public:
     V3RecursiveMutex m_mutex;  // Make sure only single thread is in class
 
@@ -361,13 +374,6 @@ public:
     int errorLimit() VL_REQUIRES(m_mutex) { return m_errorLimit; }
     void warnFatal(bool flag) VL_REQUIRES(m_mutex) { m_warnFatal = flag; }
     bool warnFatal() VL_REQUIRES(m_mutex) { return m_warnFatal; }
-    void v3errorPrep(V3ErrorCode code) VL_REQUIRES(m_mutex) {
-        m_errorStr.str("");
-        m_errorCode = code;
-        m_errorContexted = false;
-        m_errorSuppressed = false;
-    }
-    std::ostringstream& v3errorStr() VL_REQUIRES(m_mutex) { return m_errorStr; }
     V3ErrorCode errorCode() VL_REQUIRES(m_mutex) { return m_errorCode; }
     bool errorContexted() VL_REQUIRES(m_mutex) { return m_errorContexted; }
     int warnCount() VL_REQUIRES(m_mutex) { return m_warnCount; }
@@ -385,7 +391,6 @@ public:
     void describedWarnings(bool flag) VL_REQUIRES(m_mutex) { m_describedWarnings = flag; }
     int tellManual() VL_REQUIRES(m_mutex) { return m_tellManual; }
     void tellManual(int level) VL_REQUIRES(m_mutex) { m_tellManual = level; }
-    void v3errorEnd(std::ostringstream& sstr, const string& extra = "") VL_REQUIRES(m_mutex);
     void suppressThisWarning() VL_REQUIRES(m_mutex);
     string warnContextNone() VL_REQUIRES(m_mutex) {
         errorContexted(true);
@@ -512,49 +517,18 @@ public:
 
     // Internals for v3error()/v3fatal() macros only
     // Error end takes the string stream to output, be careful to seek() as needed
-    static void v3errorPrep(V3ErrorCode code) VL_MT_SAFE_EXCLUDES(s().m_mutex) {
-        const V3RecursiveLockGuard guard{s().m_mutex};
-        s().v3errorPrep(code);
-    }
-    static std::ostringstream& v3errorStr() VL_MT_SAFE_EXCLUDES(s().m_mutex) {
-        const V3RecursiveLockGuard guard{s().m_mutex};
-        return s().v3errorStr();
-    }
-    static void vlAbort();
+    static void v3errorAcquireLock() VL_ACQUIRE(s().m_mutex);
+    static std::ostringstream& v3errorPrep(V3ErrorCode code) VL_ACQUIRE(s().m_mutex);
+    static std::ostringstream& v3errorPrepFileLine(V3ErrorCode code, const char* file, int line) VL_ACQUIRE(s().m_mutex);
+    static std::ostringstream& v3errorStr() VL_REQUIRES(s().m_mutex);
     // static, but often overridden in classes.
-    static void v3errorEnd(std::ostringstream& sstr, const string& extra = "")
-        VL_MT_SAFE_EXCLUDES(s().m_mutex) VL_MT_SAFE {
-        const V3RecursiveLockGuard guard{s().m_mutex};
-        s().v3errorEnd(sstr, extra);
-    }
-    // We can't call 's().v3errorEnd' directly in 'v3ErrorEnd'/'v3errorEndFatal',
-    // due to bug in GCC (tested on 11.3.0 version with --enable-m32)
-    // causing internal error when backtrace is printed.
-    // Instead use this wrapper.
-    static void v3errorEndGuardedCall(std::ostringstream& sstr, const string& extra = "")
-        VL_REQUIRES(s().m_mutex) VL_MT_SAFE {
-        s().v3errorEnd(sstr, extra);
-    }
+    static void v3errorEnd(std::ostringstream& sstr, const string& extra = "") VL_RELEASE(s().m_mutex);
+    static void vlAbort();
 };
 
-// Global versions, so that if the class doesn't define a operator, we get the functions anyways.
-inline void v3errorEnd(std::ostringstream& sstr) VL_REQUIRES(V3Error::s().m_mutex) VL_MT_SAFE {
-    V3Error::v3errorEndGuardedCall(sstr);
-}
-inline void v3errorEndFatal(std::ostringstream& sstr)
-    VL_REQUIRES(V3Error::s().m_mutex) VL_MT_SAFE {
-    V3Error::v3errorEndGuardedCall(sstr);
-    assert(0);  // LCOV_EXCL_LINE
-    VL_UNREACHABLE;
-}
-
-#ifndef V3ERROR_NO_GLOBAL_
-#define V3ErrorLockAndCheckStopRequested \
-    V3Error::s().m_mutex.lockCheckStopRequest( \
-        []() -> void { V3ThreadPool::s().waitIfStopRequested(); })
-#else
-#define V3ErrorLockAndCheckStopRequested V3Error::s().m_mutex.lock()
-#endif
+// Global versions, so that if the class doesn't define an operator, we get the functions anyway.
+void v3errorEnd(std::ostringstream& sstr) VL_RELEASE(V3Error::s().m_mutex);
+void v3errorEndFatal(std::ostringstream& sstr) VL_RELEASE(V3Error::s().m_mutex);
 
 // Theses allow errors using << operators: v3error("foo"<<"bar");
 // Careful, you can't put () around msg, as you would in most macro definitions
@@ -565,13 +539,9 @@ inline void v3errorEndFatal(std::ostringstream& sstr)
 // but we are unlocking the mutex after function using comma operator.
 // This way macros should also work when they are in 'if' stmt without '{}'.
 #define v3warnCode(code, msg) \
-    v3errorEnd((V3ErrorLockAndCheckStopRequested, V3Error::s().v3errorPrep(code), \
-                (V3Error::s().v3errorStr() << msg), V3Error::s().v3errorStr())), \
-        V3Error::s().m_mutex.unlock()
+    v3errorEnd((V3Error::v3errorPrep(code) << msg, V3Error::v3errorStr()))
 #define v3warnCodeFatal(code, msg) \
-    v3errorEndFatal((V3ErrorLockAndCheckStopRequested, V3Error::s().v3errorPrep(code), \
-                     (V3Error::s().v3errorStr() << msg), V3Error::s().v3errorStr())), \
-        V3Error::s().m_mutex.unlock()
+    v3errorEndFatal((V3Error::v3errorPrep(code) << msg, V3Error::v3errorStr()))
 #define v3warn(code, msg) v3warnCode(V3ErrorCode::code, msg)
 #define v3info(msg) v3warnCode(V3ErrorCode::EC_INFO, msg)
 #define v3error(msg) v3warnCode(V3ErrorCode::EC_ERROR, msg)
@@ -580,14 +550,12 @@ inline void v3errorEndFatal(std::ostringstream& sstr)
 #define v3fatalExit(msg) v3warnCodeFatal(V3ErrorCode::EC_FATALEXIT, msg)
 // Use this instead of fatal() to mention the source code line.
 #define v3fatalSrc(msg) \
-    v3warnCodeFatal(V3ErrorCode::EC_FATALSRC, \
-                    __FILE__ << ":" << std::dec << __LINE__ << ": " << msg)
+    v3errorEndFatal( \
+        (V3Error::v3errorPrepFileLine(V3ErrorCode::EC_FATALSRC, __FILE__, __LINE__) << msg, \
+         V3Error::v3errorStr()))
 // Use this when normal v3fatal is called in static method that overrides fileline.
 #define v3fatalStatic(msg) \
-    (::v3errorEndFatal((V3ErrorLockAndCheckStopRequested, \
-                        V3Error::s().v3errorPrep(V3ErrorCode::EC_FATAL), \
-                        (V3Error::s().v3errorStr() << msg), V3Error::s().v3errorStr()))), \
-        V3Error::s().m_mutex.unlock()
+    ::v3errorEndFatal((V3Error::v3errorPrep(V3ErrorCode::EC_FATAL) << msg, V3Error::v3errorStr()))
 
 #define UINFO(level, stmsg) \
     do { \