--- a/src/V3Width.cpp
+++ b/src/V3Width.cpp
@@ -121,13 +121,13 @@ std::ostream& operator<<(std::ostream& str, const Castable& rhs) {
 }
 
 #define v3widthWarn(lhs, rhs, msg) \
-    v3errorEnd((V3Error::v3errorPrep((lhs) < (rhs)   ? V3ErrorCode::WIDTHTRUNC \
-                                     : (lhs) > (rhs) ? V3ErrorCode::WIDTHEXPAND \
-                                                     : V3ErrorCode::WIDTH) \
-                    << msg, \
-                V3Error::v3errorStr()))
+    v3errorEnd( \
+        v3errorBuildMessage(V3Error::v3errorPrep((lhs) < (rhs)   ? V3ErrorCode::WIDTHTRUNC \
+                                                 : (lhs) > (rhs) ? V3ErrorCode::WIDTHEXPAND \
+                                                                 : V3ErrorCode::WIDTH), \
+                            msg))
 
-//######################################################################
+// ######################################################################
 // Width state, as a visitor of each AstNode
 
 class WidthVP final {