@wsnyder 2023-08-29
So, I'm suggesting the macro calls a V3Error::getLock(). The unlocking is fine in errorEnd.