@donlon 2023-08-29
Oh, I got it. Indeed there are some function likes `nodep->warnMore()` requires the lock acquired. But it seems that `V3Error::v3errorPrep()` (LHS of the operator <<) is always called before any function in the msg list (RHS of the operator <<), however I'm not sure whether it's an UB.