@donlon 2023-08-29
There's [an article](https://www.cppstories.com/2021/evaluation-order-cpp17/) says that functions in chained << operators can be called in any order before C++17. To save another function call, I came up with two solution. The first is to use separate statements to clarify the order, but this cannot be compiled in MSVC (probably unsupported).

```C++
v3errorEnd((({std::ostringstream& os = V3Error::v3errorPrep(code); os << msg}), v3errorStr()))
```

The second is to use a self-called anonymous lambda.

```C++
v3errorEnd(
    [&]() -> std::ostringstream& {
        std::ostringstream& os = V3Error::v3errorPrep(code);
        os << msg;
        return os;
    }())
```

But the lambda cannot be inlined and make the program slightly bigger, so now it's modified to

```C++
v3errorEnd((V3Error::v3errorPrep(code), static_cast<std::ostringstream&>(V3Error::v3errorStr() << msg)))
```

It's just ok.