@wsnyder 2023-08-29
Thanks for the research. Please add a comment near that code summarizing why the comma is needed until C++17.  Then looks good to go.